+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Below, you can find all releases of this expansion and their changes.
 
+## [1.2.1]
+- Added Console logs for invalid placeholders.
+- Make provided rounding-mode lowercase to prevent issues.
+
 ## [1.2.0]
 - Restructured placeholder Syntax.  
 New Placeholders:
@@ -52,3 +56,4 @@ This version still has `1.0.7` defined as expansion version. Please update to [t
 [1.0.7]: https://api.extendedclip.com/expansions/math/versions/math-107
 [1.1.0]: https://api.extendedclip.com/expansions/math/versions/math-110
 [1.2.0]: https://api.extendedclip.com/expansions/math/versions/math-120
+[1.2.1]: https://api.extendedclip.com/expansions/math/versions/math-121